[CHORE]: DB update - Update database

+++ src/main/java/com/example/cashcow_api/dtos/milk/CustomerSaleTotalDTO.java
@@ -1,5 +1,7 @@
 package com.example.cashcow_api.dtos.milk;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 
@@ -11,11 +13,11 @@ import lombok.NoArgsConstructor;
 @JsonInclude(value = Include.NON_NULL)
 public class CustomerSaleTotalDTO {
  
-    private Double amount;
+    private BigDecimal amount;
 
     private Double quantity;
 
-    public CustomerSaleTotalDTO(Double amount, Double quantity){
+    public CustomerSaleTotalDTO(BigDecimal amount, Double quantity){
         setAmount(amount);
         setQuantity(quantity);
     }
