+++ src/main/java/com/example/cashcow_api/dtos/milk/MilkSaleTotalDTO.java
@@ -1,5 +1,6 @@
 package com.example.cashcow_api.dtos.milk;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -13,13 +14,13 @@ import lombok.NoArgsConstructor;
 @JsonInclude(value = Include.NON_NULL)
 public class MilkSaleTotalDTO {
     
-    private Double amount;
+    private BigDecimal amount;
 
     private LocalDate createdOn;
 
     private Double quantity;
 
-    public MilkSaleTotalDTO(Double amount, LocalDate createdOn, Double quantity){
+    public MilkSaleTotalDTO(BigDecimal amount, LocalDate createdOn, Double quantity){
         setAmount(amount);
         setCreatedOn(createdOn);
         setQuantity(quantity);
